### Summary

When Triton GPU IR is lowered into LLVM IR, we can make use of the constancy information about the result of the elementwise ops to deduplicate otherwise redundant computation. That is the contribution of this PR: the constancy is checked and, if possible, some of the values in LLVM IR are reused multiple times instead of computing equal values separately.

The change is beneficial for the PyTorch 2 / TorchInductor-generated Triton code, as the leftmost sub-indices extracted from the flat index by div / mod operations can be equal, given sufficiently large 2^n factor in the rightmost rightmost dimension(s). This makes the computation resulting in those sub-indices redundant. Consequently, under the necessary constancy conditions, the redundant indexing arithmetics can be deduplicated. We observe up to 29% decrease in the latency of some of our jagged tensor kernels (see the "Benchmarking" section below).

### Example

For this simple Triton kernel:

```python
@triton.jit
def triton_kernel(in_ptr, out_ptr, xnumel, XBLOCK: tl.constexpr):
    xoffset = tl.program_id(0) * XBLOCK
    xindex = xoffset + tl.arange(0, XBLOCK)[:]
    xmask = xindex < xnumel

    x1 = xindex // 256

    tmp = tl.load(in_ptr + x1, xmask)
    tl.store(out_ptr + xindex, tmp, xmask)


BLOCK_SIZE = 1024
input_t = torch.randn((1234,)).half().cuda()
output_t = torch.randn((1234, 256)).half().cuda()
grid = (triton.cdiv(output_t.numel(), BLOCK_SIZE),)
res = triton_kernel[grid](
    input_t,
    output_t,
    output_t.numel(),
    XBLOCK=BLOCK_SIZE,
    num_warps=4,
)
``` 

This is the Triton GPU IR annotated with the axis analysis info (including the constancy):

```
@triton_kernel_0d1d2de
%cst = arith.constant dense<256> : tensor<1024xi32, #triton_gpu.blocked<{sizePerThread = [8], threadsPerWarp = [32], warpsPerCTA = [4], order = [0], CTAsPerCGA = [1], CTASplitNum = [1], CTAOrder = [0]}>> => contiguity = [1], divisibility = [256], constancy = [1024], constant_value = 256
%c1024_i32 = arith.constant 1024 : i32 => contiguity = [1], divisibility = [1024], constancy = [1], constant_value = 1024
%0 = tt.get_program_id x : i32 => contiguity = [1], divisibility = [1], constancy = [1], constant_value = <none>
%1 = arith.muli %0, %c1024_i32 : i32 => contiguity = [1], divisibility = [1024], constancy = [1], constant_value = <none>
%2 = tt.make_range {end = 1024 : i32, start = 0 : i32} : tensor<1024xi32, #triton_gpu.blocked<{sizePerThread = [8], threadsPerWarp = [32], warpsPerCTA = [4], order = [0], CTAsPerCGA = [1], CTASplitNum = [1], CTAOrder = [0]}>> => contiguity = [1024], divisibility = [1073741824], constancy = [1], constant_value = <none>
%3 = tt.splat %1 : (i32) -> tensor<1024xi32, #triton_gpu.blocked<{sizePerThread = [8], threadsPerWarp = [32], warpsPerCTA = [4], order = [0], CTAsPerCGA = [1], CTASplitNum = [1], CTAOrder = [0]}>> => contiguity = [1], divisibility = [1024], constancy = [1024], constant_value = <none>
%4 = arith.addi %3, %2 : tensor<1024xi32, #triton_gpu.blocked<{sizePerThread = [8], threadsPerWarp = [32], warpsPerCTA = [4], order = [0], CTAsPerCGA = [1], CTASplitNum = [1], CTAOrder = [0]}>> => contiguity = [1024], divisibility = [1024], constancy = [1], constant_value = <none>
%5 = tt.splat %arg2 : (i32) -> tensor<1024xi32, #triton_gpu.blocked<{sizePerThread = [8], threadsPerWarp = [32], warpsPerCTA = [4], order = [0], CTAsPerCGA = [1], CTASplitNum = [1], CTAOrder = [0]}>> => contiguity = [1], divisibility = [16], constancy = [1024], constant_value = <none>
%6 = "triton_gpu.cmpi"(%4, %5) <{predicate = 2 : i64}> : (tensor<1024xi32, #triton_gpu.blocked<{sizePerThread = [8], threadsPerWarp = [32], warpsPerCTA = [4], order = [0], CTAsPerCGA = [1], CTASplitNum = [1], CTAOrder = [0]}>>, tensor<1024xi32, #triton_gpu.blocked<{sizePerThread = [8], threadsPerWarp = [32], warpsPerCTA = [4], order = [0], CTAsPerCGA = [1], CTASplitNum = [1], CTAOrder = [0]}>>) -> tensor<1024xi1, #triton_gpu.blocked<{sizePerThread = [8], threadsPerWarp = [32], warpsPerCTA = [4], order = [0], CTAsPerCGA = [1], CTASplitNum = [1], CTAOrder = [0]}>> => contiguity = [1], divisibility = [1], constancy = [16], constant_value = <none>
%7 = arith.divsi %4, %cst : tensor<1024xi32, #triton_gpu.blocked<{sizePerThread = [8], threadsPerWarp = [32], warpsPerCTA = [4], order = [0], CTAsPerCGA = [1], CTASplitNum = [1], CTAOrder = [0]}>> => contiguity = [1], divisibility = [1], constancy = [256], constant_value = <none>
%8 = tt.splat %arg0 : (!tt.ptr<f16, 1>) -> tensor<1024x!tt.ptr<f16, 1>, #triton_gpu.blocked<{sizePerThread = [8], threadsPerWarp = [32], warpsPerCTA = [4], order = [0], CTAsPerCGA = [1], CTASplitNum = [1], CTAOrder = [0]}>> => contiguity = [1], divisibility = [16], constancy = [1024], constant_value = <none>
%9 = tt.addptr %8, %7 : tensor<1024x!tt.ptr<f16, 1>, #triton_gpu.blocked<{sizePerThread = [8], threadsPerWarp = [32], warpsPerCTA = [4], order = [0], CTAsPerCGA = [1], CTASplitNum = [1], CTAOrder = [0]}>>, tensor<1024xi32, #triton_gpu.blocked<{sizePerThread = [8], threadsPerWarp = [32], warpsPerCTA = [4], order = [0], CTAsPerCGA = [1], CTASplitNum = [1], CTAOrder = [0]}>> => contiguity = [1], divisibility = [2], constancy = [256], constant_value = <none>
%10 = tt.load %9, %6 {cache = 1 : i32, evict = 1 : i32, isVolatile = false} : tensor<1024xf16, #triton_gpu.blocked<{sizePerThread = [8], threadsPerWarp = [32], warpsPerCTA = [4], order = [0], CTAsPerCGA = [1], CTASplitNum = [1], CTAOrder = [0]}>> => contiguity = [1], divisibility = [1], constancy = [16], constant_value = <none>
%11 = tt.splat %arg1 : (!tt.ptr<f16, 1>) -> tensor<1024x!tt.ptr<f16, 1>, #triton_gpu.blocked<{sizePerThread = [8], threadsPerWarp = [32], warpsPerCTA = [4], order = [0], CTAsPerCGA = [1], CTASplitNum = [1], CTAOrder = [0]}>> => contiguity = [1], divisibility = [16], constancy = [1024], constant_value = <none>
%12 = tt.addptr %11, %4 : tensor<1024x!tt.ptr<f16, 1>, #triton_gpu.blocked<{sizePerThread = [8], threadsPerWarp = [32], warpsPerCTA = [4], order = [0], CTAsPerCGA = [1], CTASplitNum = [1], CTAOrder = [0]}>>, tensor<1024xi32, #triton_gpu.blocked<{sizePerThread = [8], threadsPerWarp = [32], warpsPerCTA = [4], order = [0], CTAsPerCGA = [1], CTASplitNum = [1], CTAOrder = [0]}>> => contiguity = [1024], divisibility = [16], constancy = [1], constant_value = <none>
```

Before this PR, the LLVM IR (shown below) and the PTX (not shown here) both contain 8 redundant computations of the same `x1` value for each element in the thread; the 8 loads are performed from the 8 independent pointers containing the same address:

```
define void @triton_kernel_0d1d2de(ptr addrspace(1) %0, ptr addrspace(1) %1, i32 %2) local_unnamed_addr !dbg !5 {
  %4 = tail call i32 @llvm.nvvm.read.ptx.sreg.tid.x(), !dbg !8
  %5 = shl i32 %4, 3, !dbg !8
  %6 = and i32 %5, 1016, !dbg !8
  %7 = tail call i32 asm "mov.u32 $0, %ctaid.x;", "=r"() #1, !dbg !9
  %8 = shl i32 %7, 10, !dbg !10
  %9 = or i32 %8, %6, !dbg !11
  %10 = or i32 %9, 1, !dbg !11
  %11 = or i32 %9, 2, !dbg !11
  %12 = or i32 %9, 3, !dbg !11
  %13 = or i32 %9, 4, !dbg !11
  %14 = or i32 %9, 5, !dbg !11
  %15 = or i32 %9, 6, !dbg !11
  %16 = or i32 %9, 7, !dbg !11
  %17 = icmp slt i32 %9, %2, !dbg !12
  %18 = icmp slt i32 %10, %2, !dbg !12
  %19 = icmp slt i32 %11, %2, !dbg !12
  %20 = icmp slt i32 %12, %2, !dbg !12
  %21 = icmp slt i32 %13, %2, !dbg !12
  %22 = icmp slt i32 %14, %2, !dbg !12
  %23 = icmp slt i32 %15, %2, !dbg !12
  %24 = icmp slt i32 %16, %2, !dbg !12
  %25 = sdiv i32 %9, 256, !dbg !13
  %26 = sdiv i32 %10, 256, !dbg !13
  %27 = sdiv i32 %11, 256, !dbg !13
  %28 = sdiv i32 %12, 256, !dbg !13
  %29 = sdiv i32 %13, 256, !dbg !13
  %30 = sdiv i32 %14, 256, !dbg !13
  %31 = sdiv i32 %15, 256, !dbg !13
  %32 = sdiv i32 %16, 256, !dbg !13
  %33 = sext i32 %25 to i64, !dbg !14
  %34 = getelementptr half, ptr addrspace(1) %0, i64 %33, !dbg !14
  %35 = sext i32 %26 to i64, !dbg !14
  %36 = getelementptr half, ptr addrspace(1) %0, i64 %35, !dbg !14
  %37 = sext i32 %27 to i64, !dbg !14
  %38 = getelementptr half, ptr addrspace(1) %0, i64 %37, !dbg !14
  %39 = sext i32 %28 to i64, !dbg !14
  %40 = getelementptr half, ptr addrspace(1) %0, i64 %39, !dbg !14
  %41 = sext i32 %29 to i64, !dbg !14
  %42 = getelementptr half, ptr addrspace(1) %0, i64 %41, !dbg !14
  %43 = sext i32 %30 to i64, !dbg !14
  %44 = getelementptr half, ptr addrspace(1) %0, i64 %43, !dbg !14
  %45 = sext i32 %31 to i64, !dbg !14
  %46 = getelementptr half, ptr addrspace(1) %0, i64 %45, !dbg !14
  %47 = sext i32 %32 to i64, !dbg !14
  %48 = getelementptr half, ptr addrspace(1) %0, i64 %47, !dbg !14
  %49 = tail call i16 asm sideeffect "mov.u16 $0, 0x0;\0A\09@$2 ld.global.b16 { $0 }, [ $1 + 0 ];", "=c,l,b"(ptr addrspace(1) %34, i1 %17) #1, !dbg !15
  %50 = tail call i16 asm sideeffect "mov.u16 $0, 0x0;\0A\09@$2 ld.global.b16 { $0 }, [ $1 + 0 ];", "=c,l,b"(ptr addrspace(1) %36, i1 %18) #1, !dbg !15
  %51 = tail call i16 asm sideeffect "mov.u16 $0, 0x0;\0A\09@$2 ld.global.b16 { $0 }, [ $1 + 0 ];", "=c,l,b"(ptr addrspace(1) %38, i1 %19) #1, !dbg !15
  %52 = tail call i16 asm sideeffect "mov.u16 $0, 0x0;\0A\09@$2 ld.global.b16 { $0 }, [ $1 + 0 ];", "=c,l,b"(ptr addrspace(1) %40, i1 %20) #1, !dbg !15
  %53 = tail call i16 asm sideeffect "mov.u16 $0, 0x0;\0A\09@$2 ld.global.b16 { $0 }, [ $1 + 0 ];", "=c,l,b"(ptr addrspace(1) %42, i1 %21) #1, !dbg !15
  %54 = tail call i16 asm sideeffect "mov.u16 $0, 0x0;\0A\09@$2 ld.global.b16 { $0 }, [ $1 + 0 ];", "=c,l,b"(ptr addrspace(1) %44, i1 %22) #1, !dbg !15
  %55 = tail call i16 asm sideeffect "mov.u16 $0, 0x0;\0A\09@$2 ld.global.b16 { $0 }, [ $1 + 0 ];", "=c,l,b"(ptr addrspace(1) %46, i1 %23) #1, !dbg !15
  %56 = tail call i16 asm sideeffect "mov.u16 $0, 0x0;\0A\09@$2 ld.global.b16 { $0 }, [ $1 + 0 ];", "=c,l,b"(ptr addrspace(1) %48, i1 %24) #1, !dbg !15
  %57 = sext i32 %9 to i64, !dbg !16
  %58 = getelementptr half, ptr addrspace(1) %1, i64 %57, !dbg !16
  %59 = insertelement <2 x i16> undef, i16 %49, i64 0, !dbg !17
  %60 = insertelement <2 x i16> %59, i16 %50, i64 1, !dbg !17
  %61 = bitcast <2 x i16> %60 to i32, !dbg !17
  %62 = insertelement <2 x i16> undef, i16 %51, i64 0, !dbg !17
  %63 = insertelement <2 x i16> %62, i16 %52, i64 1, !dbg !17
  %64 = bitcast <2 x i16> %63 to i32, !dbg !17
  %65 = insertelement <2 x i16> undef, i16 %53, i64 0, !dbg !17
  %66 = insertelement <2 x i16> %65, i16 %54, i64 1, !dbg !17
  %67 = bitcast <2 x i16> %66 to i32, !dbg !17
  %68 = insertelement <2 x i16> undef, i16 %55, i64 0, !dbg !17
  %69 = insertelement <2 x i16> %68, i16 %56, i64 1, !dbg !17
  %70 = bitcast <2 x i16> %69 to i32, !dbg !17
  tail call void asm sideeffect "@$5 st.global.v4.b32 [ $4 + 0 ], { $0, $1, $2, $3 };", "r,r,r,r,l,b"(i32 %61, i32 %64, i32 %67, i32 %70, ptr addrspace(1) %58, i1 %17) #1, !dbg !17
  ret void, !dbg !18
}
```

After this PR, the redundant computations in the LLVM IR shown above are deduplicated; as a result, the 8 loads are performed from the same pointer:

```
define void @triton_kernel_0d1d2de(ptr addrspace(1) %0, ptr addrspace(1) %1, i32 %2) local_unnamed_addr !dbg !5 {
  %4 = tail call i32 @llvm.nvvm.read.ptx.sreg.tid.x(), !dbg !8
  %5 = shl i32 %4, 3, !dbg !8
  %6 = and i32 %5, 1016, !dbg !8
  %7 = tail call i32 asm "mov.u32 $0, %ctaid.x;", "=r"() #1, !dbg !9
  %8 = shl i32 %7, 10, !dbg !10
  %9 = or i32 %8, %6, !dbg !11
  %10 = icmp slt i32 %9, %2, !dbg !12
  %11 = sdiv i32 %9, 256, !dbg !13
  %12 = sext i32 %11 to i64, !dbg !14
  %13 = getelementptr half, ptr addrspace(1) %0, i64 %12, !dbg !14
  %14 = tail call i16 asm sideeffect "mov.u16 $0, 0x0;\0A\09@$2 ld.global.b16 { $0 }, [ $1 + 0 ];", "=c,l,b"(ptr addrspace(1) %13, i1 %10) #1, !dbg !15
  %15 = tail call i16 asm sideeffect "mov.u16 $0, 0x0;\0A\09@$2 ld.global.b16 { $0 }, [ $1 + 0 ];", "=c,l,b"(ptr addrspace(1) %13, i1 %10) #1, !dbg !15
  %16 = tail call i16 asm sideeffect "mov.u16 $0, 0x0;\0A\09@$2 ld.global.b16 { $0 }, [ $1 + 0 ];", "=c,l,b"(ptr addrspace(1) %13, i1 %10) #1, !dbg !15
  %17 = tail call i16 asm sideeffect "mov.u16 $0, 0x0;\0A\09@$2 ld.global.b16 { $0 }, [ $1 + 0 ];", "=c,l,b"(ptr addrspace(1) %13, i1 %10) #1, !dbg !15
  %18 = tail call i16 asm sideeffect "mov.u16 $0, 0x0;\0A\09@$2 ld.global.b16 { $0 }, [ $1 + 0 ];", "=c,l,b"(ptr addrspace(1) %13, i1 %10) #1, !dbg !15
  %19 = tail call i16 asm sideeffect "mov.u16 $0, 0x0;\0A\09@$2 ld.global.b16 { $0 }, [ $1 + 0 ];", "=c,l,b"(ptr addrspace(1) %13, i1 %10) #1, !dbg !15
  %20 = tail call i16 asm sideeffect "mov.u16 $0, 0x0;\0A\09@$2 ld.global.b16 { $0 }, [ $1 + 0 ];", "=c,l,b"(ptr addrspace(1) %13, i1 %10) #1, !dbg !15
  %21 = tail call i16 asm sideeffect "mov.u16 $0, 0x0;\0A\09@$2 ld.global.b16 { $0 }, [ $1 + 0 ];", "=c,l,b"(ptr addrspace(1) %13, i1 %10) #1, !dbg !15
  %22 = sext i32 %9 to i64, !dbg !16
  %23 = getelementptr half, ptr addrspace(1) %1, i64 %22, !dbg !16
  %24 = insertelement <2 x i16> undef, i16 %14, i64 0, !dbg !17
  %25 = insertelement <2 x i16> %24, i16 %15, i64 1, !dbg !17
  %26 = bitcast <2 x i16> %25 to i32, !dbg !17
  %27 = insertelement <2 x i16> undef, i16 %16, i64 0, !dbg !17
  %28 = insertelement <2 x i16> %27, i16 %17, i64 1, !dbg !17
  %29 = bitcast <2 x i16> %28 to i32, !dbg !17
  %30 = insertelement <2 x i16> undef, i16 %18, i64 0, !dbg !17
  %31 = insertelement <2 x i16> %30, i16 %19, i64 1, !dbg !17
  %32 = bitcast <2 x i16> %31 to i32, !dbg !17
  %33 = insertelement <2 x i16> undef, i16 %20, i64 0, !dbg !17
  %34 = insertelement <2 x i16> %33, i16 %21, i64 1, !dbg !17
  %35 = bitcast <2 x i16> %34 to i32, !dbg !17
  tail call void asm sideeffect "@$5 st.global.v4.b32 [ $4 + 0 ], { $0, $1, $2, $3 };", "r,r,r,r,l,b"(i32 %26, i32 %29, i32 %32, i32 %35, ptr addrspace(1) %23, i1 %10) #1, !dbg !17
  ret void, !dbg !18
}
```

### Benchmarking

This is the `jagged_to_padded_dense` kernel generated by TorchInductor. The kernel is converting a jagged tensor input (represented by separate `values` and `offsets` tensors) into the corresponding dense (padded) tensor `output`, using the scalar value `3.14` for the elements not present in the sparse jagged input representation:

```python
@triton.jit
def triton_kernel(values, offsets, output, batch_size, xnumel, XBLOCK: tl.constexpr):
    xoffset = tl.program_id(0) * XBLOCK
    xindex = xoffset + tl.arange(0, XBLOCK)[:]
    xmask = xindex < xnumel

    x1 = xindex // 66560
    x2 = (xindex // 256) % 260
    x3 = xindex % 66560

    tmp0 = tl.load(offsets + (x1), xmask, eviction_policy="evict_last")
    tmp2 = tl.load(offsets + (1 + x1), xmask, eviction_policy="evict_last")
    tmp1 = tl.where(tmp0 < 0, tmp0 + 1 + batch_size, tmp0)
    tmp3 = tmp1 + x2
    tmp4 = tmp3 < tmp2
    tmp5 = tl.load(values + (x3 + (256 * tmp1)), tmp4 & xmask, other=0).to(tl.float32)
    tmp6 = tl.where(tmp4, tmp5, 3.14)

    tl.store(output + xindex, tmp6, xmask)
```

Below are the kernel runtimes before and after this PR as measured on different load factors of the jagged tensor input: i.e., the fraction of the data represented by the sparse jagged input (in other words, the average sequence length as a fraction of the maximum sequence length):

| load factor  | before this PR, ms | after this PR, ms | perf gain |
|---|---|---|---|
| 0.1 | 0.151 | 0.106 | 29% |
| 0.2 | 0.151 | 0.117 | 22% |
| 0.3 | 0.150 | 0.126 | 16% |
| 0.4 | 0.150 | 0.133 | 11% |
| 0.5 | 0.151 | 0.141 | 7% |
| 0.6 | 0.152 | 0.147 | 4% |
| 0.7 | 0.153 | 0.151 | 2% |
| 0.8 | 0.155 | 0.154 | 1% |
| 0.9 | 0.157 | 0.156 | 1% |
| 1.0 | 0.158 | 0.159 | 0% |

As it can be seen, on low load factors the heuristic introduced by this PR can lead to a performance gain of up to 29%. On high load factors, the memory I/O becomes the performance bottleneck making the index arithmetics deduplication less important.

### Test Plan

New `lit`-based unit tests have been added to `test/Conversion/dedup-by-constancy.mlir`:

```
$ lit test
-- Testing: 39 tests, 39 workers --
...

Testing Time: 0.70s
  Passed           : 38
  Expectedly Failed:  1
```