@Jokeren 2023-10-19
> @aakhundov This would benefit my kernels a lot as well. Related to this, would it be possible to achieve the same deduplication if you had non-contiguous repeating elements? For example, withx1 = xindex % 256 instead of x1 = xindex // 256.

I think the heuristic would be general since he checks the constancy of the result