@Jokeren 2023-10-19
Do you have any benchmark that demonstrate the speedup? That would be great