@aakhundov 2023-10-19
@francescocarzaniga

> For example, withx1 = xindex % 256 instead of x1 = xindex // 256.

In the case of `x1 = xindex % 256`, the resulting `x1` would have contiguity but not constancy. As the heuristic relies on the constancy for deduplication across the multiple (contiguous) elements in a single thread, I believe, dedup won't happen here.