@Jokeren 2023-10-19
@aakhundov `python3 -m pre_commit run --all-files --verbose`