@Jokeren 2023-10-19
> @aakhundov I see, that's unfortunate. It would be great to have both kinds of deduplication working as they are complementary, but I'm no compiler expert so I have no idea how hard it would be to implement it.

Contiguity wouldn't cause duplication I think, but `x % 256` does cause duplication if x is a vector and has constancy > 1 (e.g., [4, 4, 4, 4] % 256)?