@francescocarzaniga 2023-10-19
> Contiguity wouldn't cause duplication I think, but `x % 256` does cause duplication if x is a vector and has constancy > 1 (e.g., [4, 4, 4, 4] % 256)?

Indeed in this case it would. What I am talking about is something like `[0, 1, 2, 3] % 2 -> [0, 1, 0, 1]` (this is complementary to `[0, 1, 2, 3] // 2 -> [0, 0, 1, 1]`). There is clearly some duplication going on here, but I assume it cannot be deduplicated in the same way.