@Jokeren 2023-10-19
> Indeed in this case it would. What I am talking about is something like [0, 1, 2, 3] % 2 -> [0, 1, 0, 1] (this is complementary to [0, 1, 2, 3] // 2 -> [0, 0, 1, 1]). There is clearly some duplication going on here, but I assume it cannot be deduplicated in the same way.

There's no guarantee that [2,3] exist in the same thread as [0,1]. So in your case, maybe it is more of [0, 1, ... 512] % 256? So probably there's no duplication within the same thread?

There has been a long line of architecture research that automatically discovers duplication within and across threads and design function units to address the issues. The uniform register on nvidia GPUs is one way that addresses the duplicate values across threads within the same warp.
From the compiler perspective, I think we could gather more use cases and dive into the problems.