@aakhundov 2023-10-19
> @aakhundov `python3 -m pre_commit run --all-files --verbose`

@Jokeren Fixed the linter. Could you approve the CI runs again? Thanks!