@aakhundov 2023-10-20
@Jokeren apparently, it's generally unsafe to reuse LLVM `Value`s across different `sizePerThread`-sized blocks when `elemsPerThread` is larger. A few unit tests were failing due to this (e.g., `test_masked_load`), as some essential instructions got wrongly optimized out. I've added an additional condition for that and fixed the failing unit tests that I could verify on A100. Could you approve the CI again? Thanks!