@francescocarzaniga 2023-10-20
> There's no guarantee that [2,3] exist in the same thread as [0,1]. So in your case, maybe it is more of [0, 1, ... 512] % 256? So probably there's no duplication within the same thread?
> 
> There has been a long line of architecture research that automatically discovers duplication within and across threads and design function units to address the issues. The uniform register on nvidia GPUs is one way that addresses the duplicate values across threads within the same warp. From the compiler perspective, I think we could gather more use cases and dive into the problems.

My case is actually something like `[0, ..., 2^N] // M` (where M is not necessarily a power of 2), so there might be duplication in the same thread. I see what you mean however, that's why I was worried that optimising this case would not be as straightforward, but it would still be very useful.