@ThomasRaoux 2023-10-20
so if I understand the optimization correctly the idea is that:
```
%10 = or i32 %9, 1, !dbg !11
%25 = sdiv i32 %10, 256, !dbg !13
```
can be changed to:
```
%25 = sdiv i32 %9, 256, !dbg !13
```

In general this is not correct for negative integers however this would be the case for unsigned integers (https://godbolt.org/z/7EqzzKEbM). Maybe the solution is to just cast this to an unsigned integer so that llvm will optimize things out. Technically llvm could also figure it out from the range but this would be more complicated.