@aakhundov 2023-10-20
> so if I understand the optimization correctly the idea is that:
> 
> ```
> %10 = or i32 %9, 1, !dbg !11
> %25 = sdiv i32 %10, 256, !dbg !13
> ```
> 
> can be changed to:
> 
> ```
> %25 = sdiv i32 %9, 256, !dbg !13
> ```
> 
> In general this is not correct for negative integers however this would be the case for unsigned integers (https://godbolt.org/z/7EqzzKEbM). Maybe the solution is to just cast this to an unsigned integer so that llvm will optimize things out. Technically llvm could also figure it out from the range but this would be more complicated.

@ThomasRaoux I don't think that `or i32` is completely eliminated. From what I see in the example in the PR description, the following fragment with the 8 instances of each instruction in the original PTX:

```
  %9 = or i32 %8, %6, !dbg !11
  %10 = or i32 %9, 1, !dbg !11
  %11 = or i32 %9, 2, !dbg !11
  %12 = or i32 %9, 3, !dbg !11
  %13 = or i32 %9, 4, !dbg !11
  %14 = or i32 %9, 5, !dbg !11
  %15 = or i32 %9, 6, !dbg !11
  %16 = or i32 %9, 7, !dbg !11
  %17 = icmp slt i32 %9, %2, !dbg !12
  %18 = icmp slt i32 %10, %2, !dbg !12
  %19 = icmp slt i32 %11, %2, !dbg !12
  %20 = icmp slt i32 %12, %2, !dbg !12
  %21 = icmp slt i32 %13, %2, !dbg !12
  %22 = icmp slt i32 %14, %2, !dbg !12
  %23 = icmp slt i32 %15, %2, !dbg !12
  %24 = icmp slt i32 %16, %2, !dbg !12
  %25 = sdiv i32 %9, 256, !dbg !13
  %26 = sdiv i32 %10, 256, !dbg !13
  %27 = sdiv i32 %11, 256, !dbg !13
  %28 = sdiv i32 %12, 256, !dbg !13
  %29 = sdiv i32 %13, 256, !dbg !13
  %30 = sdiv i32 %14, 256, !dbg !13
  %31 = sdiv i32 %15, 256, !dbg !13
  %32 = sdiv i32 %16, 256, !dbg !13
  %33 = sext i32 %25 to i64, !dbg !14
  %34 = getelementptr half, ptr addrspace(1) %0, i64 %33, !dbg !14
  %35 = sext i32 %26 to i64, !dbg !14
  %36 = getelementptr half, ptr addrspace(1) %0, i64 %35, !dbg !14
  %37 = sext i32 %27 to i64, !dbg !14
  %38 = getelementptr half, ptr addrspace(1) %0, i64 %37, !dbg !14
  %39 = sext i32 %28 to i64, !dbg !14
  %40 = getelementptr half, ptr addrspace(1) %0, i64 %39, !dbg !14
  %41 = sext i32 %29 to i64, !dbg !14
  %42 = getelementptr half, ptr addrspace(1) %0, i64 %41, !dbg !14
  %43 = sext i32 %30 to i64, !dbg !14
  %44 = getelementptr half, ptr addrspace(1) %0, i64 %43, !dbg !14
  %45 = sext i32 %31 to i64, !dbg !14
  %46 = getelementptr half, ptr addrspace(1) %0, i64 %45, !dbg !14
  %47 = sext i32 %32 to i64, !dbg !14
  %48 = getelementptr half, ptr addrspace(1) %0, i64 %47, !dbg !14
```

gets deduplicated into a single instance of each instruction:

```
  %9 = or i32 %8, %6, !dbg !11
  %10 = icmp slt i32 %9, %2, !dbg !12
  %11 = sdiv i32 %9, 256, !dbg !13
  %12 = sext i32 %11 to i64, !dbg !14
  %13 = getelementptr half, ptr addrspace(1) %0, i64 %12, !dbg !14
```

as the result of each of those operations (or the downstream ones) has constancy covering the whole 8 values in the original PTX. Does this make sense?